refactor: Remote source populate data
