+++ app/models/expenditure.rb
@@ -1,6 +1,6 @@
 class Expenditure < ApplicationRecord
   RECEIPT_TYPE = %w[INVOICE RECEIPT EXPENSES_ABROAD NO_SPECIFIED].freeze
-  enum receipt_type: %i[INVOICE RECEIPT EXPENSES_ABROAD NO_SPECIFIED]
+  enum receipt_type: %i[INVOICE RECEIPT NO_SPECIFIED NO_SPECIFIED2 EXPENSES_ABROAD]
 
   belongs_to :deputy
   belongs_to :category
@@ -9,5 +9,4 @@ class Expenditure < ApplicationRecord
   validates_associated :category
   validates :receipt_type, inclusion: { in: RECEIPT_TYPE }, presence: true
   validates :period, presence: true
-
 end
